+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Added a validator for README.md files.
 * Fixed an issue where the validator of reputation(Indicator Type) did not check on the details field.
 * Removed the *yml_type* argument in the *split-yml* command.
+* Fixed the --insecure argument in *upload*
+* Added the --insecure argument in *run-playbook*
 
 ### 0.4.3
 * Fixed an issue where the incident and indicator field BC check failed.